rename: #1 CLIをインターフェースアダプタと捉えて，ディレクトリ名変更
